refactor(mcp): remove hub tools from gateway
