+++ docs/notes/opentelemetry.md
@@ -129,12 +129,12 @@ To ensure dashboard compatibility across the entire fleet, all signals follow th
 - `mcp.tool.<name>`: Root Span for each tool execution (e.g., `mcp.tool.inspect_pods`)
 - **Attributes:**
   - `mcp.tool`: The specific tool name.
-  - `mcp.service`: The domain discriminator (`mcp.telemetry`, `mcp.pods`, or `mcp.hub`).
+  - `mcp.service`: The domain discriminator (`mcp.telemetry`, `mcp.pods`, or `mcp.network`).
 
 **Logging Pattern:**
 The unified agent emits logs for sequential provider initialization and execution visibility:
 
-- `registered hub tools (mcp.hub)`, `registered pods tools (mcp.pods)`, `registered telemetry tools (mcp.telemetry)`
+- `registered network tools`, `registered pods tools (mcp.pods)`, `registered telemetry tools (mcp.telemetry)`
 - `executing PromQL query`, `executing LogQL query`, `retrieving trace from Tempo`
 - `investigating incident` (Macro-tool orchestration)
 - `using local kubeconfig` (Diagnostic logs for the Pods provider)
