+++ internal/mcp/registry.go
@@ -2,12 +2,12 @@ package mcp
 
 import (
 	"context"
-	"encoding/json"
+	"fmt"
 
 	"github.com/modelcontextprotocol/go-sdk/mcp"
 
 	"observability-hub/internal/mcp/providers"
-	"observability-hub/internal/mcp/tools/hub"
+	"observability-hub/internal/mcp/tools/network"
 	"observability-hub/internal/mcp/tools/pods"
 	"observability-hub/internal/mcp/tools/telemetry"
 	libtelemetry "observability-hub/internal/telemetry"
@@ -42,58 +42,22 @@ func RegisterTelemetryTools(server *mcp.Server, provider *providers.TelemetryPro
 
 func handleQueryMetrics(provider *providers.TelemetryProvider, serviceName string) mcp.ToolHandlerFor[telemetry.QueryMetricsInput, any] {
 	handler := telemetry.NewQueryMetricsHandler(provider.QueryMetrics)
-	return InstrumentHandler("query_metrics", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input telemetry.QueryMetricsInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
+	return NewJSONToolHandler("query_metrics", serviceName, handler.Execute)
 }
 
 func handleQueryLogs(provider *providers.TelemetryProvider, serviceName string) mcp.ToolHandlerFor[telemetry.QueryLogsInput, any] {
 	handler := telemetry.NewQueryLogsHandler(provider.QueryLogs)
-	return InstrumentHandler("query_logs", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input telemetry.QueryLogsInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
+	return NewJSONToolHandler("query_logs", serviceName, handler.Execute)
 }
 
 func handleInvestigateIncident(provider *providers.TelemetryProvider, serviceName string) mcp.ToolHandlerFor[telemetry.InvestigateIncidentInput, any] {
 	handler := telemetry.NewInvestigateIncidentHandler(provider.QueryMetrics, provider.QueryLogs, provider.QueryTraces)
-	return InstrumentHandler("investigate_incident", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input telemetry.InvestigateIncidentInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
+	return NewJSONToolHandler("investigate_incident", serviceName, handler.Execute)
 }
 
 func handleQueryTraces(provider *providers.TelemetryProvider, serviceName string) mcp.ToolHandlerFor[telemetry.QueryTracesInput, any] {
 	handler := telemetry.NewQueryTracesHandler(provider.QueryTraces)
-	return InstrumentHandler("query_traces", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input telemetry.QueryTracesInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
+	return NewJSONToolHandler("query_traces", serviceName, handler.Execute)
 }
 
 // --- Pods Tools ---
@@ -130,159 +94,43 @@ func RegisterPodsTools(server *mcp.Server, provider *providers.PodsProvider, ser
 
 func handleInspectPods(provider *providers.PodsProvider, serviceName string) mcp.ToolHandlerFor[pods.PodsInput, any] {
 	handler := pods.NewInspectPodsHandler(provider.ListPods)
-	return InstrumentHandler("inspect_pods", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input pods.PodsInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
+	return NewJSONToolHandler("inspect_pods", serviceName, handler.Execute)
 }
 
 func handleDescribePod(provider *providers.PodsProvider, serviceName string) mcp.ToolHandlerFor[pods.PodsInput, any] {
 	handler := pods.NewDescribePodHandler(provider.GetPod)
-	return InstrumentHandler("describe_pod", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input pods.PodsInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
+	return NewJSONToolHandler("describe_pod", serviceName, handler.Execute)
 }
 
 func handleListPodEvents(provider *providers.PodsProvider, serviceName string) mcp.ToolHandlerFor[pods.PodsInput, any] {
 	handler := pods.NewListPodEventsHandler(provider.ListEvents)
-	return InstrumentHandler("list_pod_events", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input pods.PodsInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
+	return NewJSONToolHandler("list_pod_events", serviceName, handler.Execute)
 }
 
 func handleGetPodLogs(provider *providers.PodsProvider, serviceName string) mcp.ToolHandlerFor[pods.PodLogsInput, any] {
 	handler := pods.NewGetPodLogsHandler(provider.GetPodLogs)
-	return InstrumentHandler("get_pod_logs", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input pods.PodLogsInput) (*mcp.CallToolResult, any, error) {
+	return NewTextToolHandler("get_pod_logs", serviceName, func(ctx context.Context, input pods.PodLogsInput) (string, error) {
 		result, err := handler.Execute(ctx, input)
 		if err != nil {
-			return nil, nil, err
+			return "", err
 		}
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: result.(string)}},
-		}, nil, nil
+		text, ok := result.(string)
+		if !ok {
+			return "", fmt.Errorf("get_pod_logs returned %T, want string", result)
+		}
+		return text, nil
 	})
 }
 
 func handleDeletePod(provider *providers.PodsProvider, serviceName string) mcp.ToolHandlerFor[pods.DeletePodInput, any] {
 	handler := pods.NewDeletePodHandler(provider.DeletePod)
-	return InstrumentHandler("delete_pod", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input pods.DeletePodInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
-}
-
-// --- Hub Tools ---
-
-// RegisterHubTools registers all host-level and platform status tools to the MCP server.
-func RegisterHubTools(server *mcp.Server, provider *providers.HubProvider, serviceName string) {
-	mcp.AddTool(server, &mcp.Tool{
-		Name:        "hub_inspect_platform",
-		Description: "Get an executive summary of the entire platform health (See skills/platform/SKILL.md for guidance)",
-	}, handleInspectPlatform(provider, serviceName))
-
-	mcp.AddTool(server, &mcp.Tool{
-		Name:        "hub_inspect_host",
-		Description: "Inspect physical resource pressure (Load, Memory, Disk) on the main server (See skills/host/SKILL.md for guidance)",
-	}, handleInspectHost(provider, serviceName))
-
-	mcp.AddTool(server, &mcp.Tool{
-		Name:        "hub_list_host_services",
-		Description: "List and check status of core systemd units (ingestion, proxy, openbao) (See skills/host/SKILL.md for guidance)",
-	}, handleListHostServices(provider, serviceName))
-
-	mcp.AddTool(server, &mcp.Tool{
-		Name:        "hub_query_service_logs",
-		Description: "Query systemd journal logs for a specific service since a relative time (e.g., past 5m, 1h) (See skills/host/SKILL.md for guidance)",
-	}, handleQueryServiceLogs(provider, serviceName))
-
-	libtelemetry.Info("registered hub tools", "count", 4)
-}
-
-func handleInspectPlatform(provider *providers.HubProvider, serviceName string) mcp.ToolHandlerFor[hub.HubInput, any] {
-	handler := hub.NewInspectPlatformHandler(provider.InspectPlatform)
-	return InstrumentHandler("hub_inspect_platform", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input hub.HubInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
-}
-
-func handleInspectHost(provider *providers.HubProvider, serviceName string) mcp.ToolHandlerFor[hub.HubInput, any] {
-	handler := hub.NewInspectHostHandler(provider.InspectHost)
-	return InstrumentHandler("hub_inspect_host", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input hub.HubInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
-}
-
-func handleListHostServices(provider *providers.HubProvider, serviceName string) mcp.ToolHandlerFor[hub.HubInput, any] {
-	handler := hub.NewListHostServicesHandler(provider.ListHostServices)
-	return InstrumentHandler("hub_list_host_services", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input hub.HubInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		text, _ := json.Marshal(result)
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: string(text)}},
-		}, nil, nil
-	})
-}
-
-func handleQueryServiceLogs(provider *providers.HubProvider, serviceName string) mcp.ToolHandlerFor[hub.HubInput, any] {
-	handler := hub.NewQueryServiceLogsHandler(provider.QueryServiceLogs)
-	return InstrumentHandler("hub_query_service_logs", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input hub.HubInput) (*mcp.CallToolResult, any, error) {
-		result, err := handler.Execute(ctx, input)
-		if err != nil {
-			return nil, nil, err
-		}
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: result.(string)}},
-		}, nil, nil
-	})
+	return NewJSONToolHandler("delete_pod", serviceName, handler.Execute)
 }
 
 // --- Network Tools ---
 
 // RegisterNetworkTools registers all networking-related tools (Hubble) to the MCP server.
-func RegisterNetworkTools(server *mcp.Server, provider *providers.HubProvider, serviceName string) {
+func RegisterNetworkTools(server *mcp.Server, provider *providers.NetworkProvider, serviceName string) {
 	mcp.AddTool(server, &mcp.Tool{
 		Name:        "observe_network_flows",
 		Description: "Query real-time network flows from Hubble Relay (See skills/network/SKILL.md for guidance)",
@@ -291,15 +139,17 @@ func RegisterNetworkTools(server *mcp.Server, provider *providers.HubProvider, s
 	libtelemetry.Info("registered network tools", "count", 1)
 }
 
-func handleObserveNetworkFlows(provider *providers.HubProvider, serviceName string) mcp.ToolHandlerFor[hub.ObserveNetworkFlowsInput, any] {
-	handler := hub.NewObserveNetworkFlowsHandler(provider.QueryHubbleFlows)
-	return InstrumentHandler("observe_network_flows", serviceName, func(ctx context.Context, _ *mcp.CallToolRequest, input hub.ObserveNetworkFlowsInput) (*mcp.CallToolResult, any, error) {
+func handleObserveNetworkFlows(provider *providers.NetworkProvider, serviceName string) mcp.ToolHandlerFor[network.ObserveNetworkFlowsInput, any] {
+	handler := network.NewObserveNetworkFlowsHandler(provider.QueryHubbleFlows)
+	return NewTextToolHandler("observe_network_flows", serviceName, func(ctx context.Context, input network.ObserveNetworkFlowsInput) (string, error) {
 		result, err := handler.Execute(ctx, input)
 		if err != nil {
-			return nil, nil, err
+			return "", err
 		}
-		return &mcp.CallToolResult{
-			Content: []mcp.Content{&mcp.TextContent{Text: result.(string)}},
-		}, nil, nil
+		text, ok := result.(string)
+		if !ok {
+			return "", fmt.Errorf("observe_network_flows returned %T, want string", result)
+		}
+		return text, nil
 	})
 }
