--- skills/platform/references/api-specs.md
@@ -1,11 +1,0 @@
-# Platform API Specifications
-
-Detailed input schemas for platform-level MCP tools.
-
-## Tools
-
-### hub_inspect_platform
-
-- **Input:**
-  - `(none)` (Empty object `{}`).
-- **Returns:** An executive summary of platform-wide health status.
